// canvas-modern-ui/packages/lti-service/src/lib/config.d.ts
import { z } from 'zod';
import 'dotenv/config';
declare const configSchema: z.ZodObject<{
    PORT: z.ZodDefault<z.ZodNumber>;
    NODE_ENV: z.ZodDefault<z.ZodEnum<["development", "production", "test"]>>;
    LOG_LEVEL: z.ZodDefault<z.ZodEnum<["trace", "debug", "info", "warn", "error"]>>;
    SESSION_SECRET: z.ZodString;
    LTI_TOOL_KID: z.ZodDefault<z.ZodString>;
    LTI_TOOL_PRIVATE_KEY_PEM: z.ZodString;
    LTI_ISSUER: z.ZodString;
    LTI_CLIENT_ID: z.ZodString;
    LTI_AUTHORIZATION_ENDPOINT: z.ZodString;
    LTI_TOKEN_ENDPOINT: z.ZodString;
    LTI_JWKS_ENDPOINT: z.ZodString;
    BASE_URL: z.ZodString;
    LTI_INITIATE_LOGIN_URI: z.ZodString;
    LTI_REDIRECT_URI: z.ZodString;
    MODERN_UI_URL: z.ZodDefault<z.ZodString>;
    CANVAS_API_BASE_URL: z.ZodOptional<z.ZodString>;
    CANVAS_OAUTH_CLIENT_ID: z.ZodOptional<z.ZodString>;
    CANVAS_OAUTH_CLIENT_SECRET: z.ZodOptional<z.ZodString>;
    SENTRY_DSN: z.ZodUnion<[z.ZodOptional<z.ZodString>, z.ZodLiteral<"">]>;
}, "strip", z.ZodTypeAny, {
    PORT: number;
    NODE_ENV: "development" | "production" | "test";
    LOG_LEVEL: "info" | "warn" | "error" | "trace" | "debug";
    SESSION_SECRET: string;
    LTI_TOOL_KID: string;
    LTI_TOOL_PRIVATE_KEY_PEM: string;
    LTI_ISSUER: string;
    LTI_CLIENT_ID: string;
    LTI_AUTHORIZATION_ENDPOINT: string;
    LTI_TOKEN_ENDPOINT: string;
    LTI_JWKS_ENDPOINT: string;
    BASE_URL: string;
    LTI_INITIATE_LOGIN_URI: string;
    LTI_REDIRECT_URI: string;
    MODERN_UI_URL: string;
    CANVAS_API_BASE_URL?: string | undefined;
    CANVAS_OAUTH_CLIENT_ID?: string | undefined;
    CANVAS_OAUTH_CLIENT_SECRET?: string | undefined;
    SENTRY_DSN?: string | undefined;
}, {
    SESSION_SECRET: string;
    LTI_TOOL_PRIVATE_KEY_PEM: string;
    LTI_ISSUER: string;
    LTI_CLIENT_ID: string;
    LTI_AUTHORIZATION_ENDPOINT: string;
    LTI_TOKEN_ENDPOINT: string;
    LTI_JWKS_ENDPOINT: string;
    BASE_URL: string;
    LTI_INITIATE_LOGIN_URI: string;
    LTI_REDIRECT_URI: string;
    PORT?: number | undefined;
    NODE_ENV?: "development" | "production" | "test" | undefined;
    LOG_LEVEL?: "info" | "warn" | "error" | "trace" | "debug" | undefined;
    LTI_TOOL_KID?: string | undefined;
    MODERN_UI_URL?: string | undefined;
    CANVAS_API_BASE_URL?: string | undefined;
    CANVAS_OAUTH_CLIENT_ID?: string | undefined;
    CANVAS_OAUTH_CLIENT_SECRET?: string | undefined;
    SENTRY_DSN?: string | undefined;
}>;
export type Config = z.infer<typeof configSchema>;
export declare const config: {
    PORT: number;
    NODE_ENV: "development" | "production" | "test";
    LOG_LEVEL: "info" | "warn" | "error" | "trace" | "debug";
    SESSION_SECRET: string;
    LTI_TOOL_KID: string;
    LTI_TOOL_PRIVATE_KEY_PEM: string;
    LTI_ISSUER: string;
    LTI_CLIENT_ID: string;
    LTI_AUTHORIZATION_ENDPOINT: string;
    LTI_TOKEN_ENDPOINT: string;
    LTI_JWKS_ENDPOINT: string;
    BASE_URL: string;
    LTI_INITIATE_LOGIN_URI: string;
    LTI_REDIRECT_URI: string;
    MODERN_UI_URL: string;
    CANVAS_API_BASE_URL?: string | undefined;
    CANVAS_OAUTH_CLIENT_ID?: string | undefined;
    CANVAS_OAUTH_CLIENT_SECRET?: string | undefined;
    SENTRY_DSN?: string | undefined;
};
export {};
//# sourceMappingURL=config.d.ts.map